ProyectoSena/requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.36.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
et-xmlfile==1.0.1
idna==2.6
jdcal==1.4
mysqlclient==1.3.12
oauthlib==2.1.0
openpyxl==2.5.3
Pillow==5.1.0
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
